requirements.txt
django
djangorestframework
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
django-filter

